Refactor and clean up testWrapping
