+++ OpenSim/Tests/Wrapping/CMakeLists.txt
@@ -1,5 +1,7 @@
 
-file(GLOB TEST_PROGS "test*.cpp")
+file(GLOB TEST_PROGS "test*.cpp"
+                     # "sandbox*.cpp"
+    )
 file(GLOB TEST_FILES *.osim *.xml *.sto *.mot)
 
 OpenSimAddTests(
